requirements.txt
torchvision==0.9.0+cu111
numpy==1.16.4
torch==1.8.0+cu111
tqdm==4.59.0
pandas==0.24.2
pillow==7.1.2
h5py==2.9.0
scipy==1.2.1